[TH2-2895] Moved router out of its own folder to other abstract classes.
