+++ th2_common/schema/message/impl/rabbitmq/abstract_rabbit_batch_message_router.py
