+++ th2_common/schema/message/impl/rabbitmq/raw/rabbit_raw_batch_router.py
@@ -12,9 +12,8 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-from th2_grpc_common.common_pb2 import RawMessageBatch, RawMessage, AnyMessage, MessageGroup, MessageGroupBatch
+from th2_grpc_common.common_pb2 import AnyMessage, MessageGroup, MessageGroupBatch
 
-from th2_common.schema.filter.strategy.impl.default_filter_strategy import DefaultFilterStrategy
 from th2_common.schema.message.configuration.message_configuration import QueueConfiguration
 from th2_common.schema.message.impl.rabbitmq.configuration.subscribe_target import SubscribeTarget
 from th2_common.schema.message.impl.rabbitmq.connection.connection_manager import ConnectionManager
@@ -22,12 +21,9 @@ from th2_common.schema.message.impl.rabbitmq.group.rabbit_message_group_batch_ro
     RabbitMessageGroupBatchRouter
 from th2_common.schema.message.impl.rabbitmq.raw.rabbit_raw_batch_sender import RabbitRawBatchSender
 from th2_common.schema.message.impl.rabbitmq.raw.rabbit_raw_batch_subscriber import RabbitRawBatchSubscriber
-from th2_common.schema.message.impl.rabbitmq.router.abstract_rabbit_batch_message_router import \
-    AbstractRabbitBatchMessageRouter
 from th2_common.schema.message.message_sender import MessageSender
 from th2_common.schema.message.message_subscriber import MessageSubscriber
 from th2_common.schema.message.queue_attribute import QueueAttribute
-from th2_common.schema.util.util import get_session_alias_and_direction
 
 
 class RabbitRawBatchRouter(RabbitMessageGroupBatchRouter):
